**Lease** 1. Create Lease 2. Add tenant 3. get lease 4. remove tenant 5. Collect rent

+++ property_manager_be/src/main/java/com/property_manager/property_manager_be/repository/TenantRepository.java
@@ -3,4 +3,4 @@ package com.property_manager.property_manager_be.repository;
 import com.property_manager.property_manager_be.models.Tenant;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TenantRepository extends CrudRepository<Tenant, Integer> {}
+public interface TenantRepository extends CrudRepository<Tenant, Long> {}
